Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/xhr/open-url-redirected-sharedworker-origin-expected.txt

CONSOLE MESSAGE: line 10: ReferenceError: Can't find variable: SharedWorker

FAIL XMLHttpRequest: redirected classic shared worker scripts, origin and referrer ReferenceError: Can't find variable: SharedWorker
